requirements.txt
Django==6.0
djangorestframework==3.16.1
Pillow==12.0.0
django-cors-headers==4.9.0
